+++ 1st-gen/packages/illustrated-message/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 1.11.2
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @spectrum-web-components/base@1.11.2
+    - @spectrum-web-components/styles@1.11.2
+
+## 1.11.1
+
+### Patch Changes
+
+- Updated dependencies [[`95e1c25`](https://github.com/adobe/spectrum-web-components/commit/95e1c25672f62f3723dfa66129ae5ecdeabe578a)]:
+    - @spectrum-web-components/base@1.11.1
+    - @spectrum-web-components/styles@1.11.1
+
+## 1.11.0
+
+### Patch Changes
+
+- Updated dependencies [[`9cb816b`](https://github.com/adobe/spectrum-web-components/commit/9cb816b5ac80387fdc9bb87381f5149fecb1b595), [`7c26e3a`](https://github.com/adobe/spectrum-web-components/commit/7c26e3afa534342f09d956c7f1d1acda7e0da939)]:
+    - @spectrum-web-components/base@1.11.0
+    - @spectrum-web-components/styles@1.11.0
+
 ## 1.10.0
 
 ### Patch Changes